backend/requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
pandas==2.3.2
PyPDF2==3.0.1
requests==2.32.5
python-dotenv==1.1.1
python-pptx==0.6.21
openpyxl==3.1.5
